test/refactor: move RPC fuzz helpers to util/rpc
Moved ConsumeScalarRPCArgument to its own fuzz util file
