+++ src/test/fuzz/util/CMakeLists.txt
@@ -7,6 +7,7 @@ add_library(test_fuzz STATIC EXCLUDE_FROM_ALL
   descriptor.cpp
   mempool.cpp
   net.cpp
+  rpc.cpp
   threadinterrupt.cpp
   ../fuzz.cpp
   ../util.cpp
